Dynamically resource allocation
The memory and time for the nextflow jobs will be dynamically allocated. If some of the jobs failed at the first try, the memory and time for the second try will increase. 

Errors other than the "140" (not enough time or memory) will be ignored, the overall workflow execution will not be stopped. 

Maximum retry times for a job is set at 2, but can be increased as needed. We can create a flag from the outside to control this nextflow param.